# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Gymie MVP API
  version: 1.0.0
  description: API spec for Gymie MVP backend (running + lifting + AI plan generation).
servers:
  - url: http://localhost:4101/api
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Error409'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Error401'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/logout:
    delete:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '204':
          description: Logged out
  /users/me:
    get:
      tags: [Users]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePatch'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
  /users/me/preferences:
    put:
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencePatch'
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            details:
              nullable: true
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        age: { type: integer, nullable: true }
        gender: { type: string, nullable: true }
        height_cm: { type: number, nullable: true }
        weight_kg: { type: number, nullable: true }
        goal: { type: string, nullable: true }
        fitness_level: { type: string, nullable: true }
        unit_system: { type: string }
        training_type: { type: string }
        available_days:
          type: array
          items: { type: integer }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1, maxLength: 50 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        age: { type: integer }
        gender: { type: string, enum: [male, female, other] }
        height_cm: { type: number }
        weight_kg: { type: number }
        goal: { type: string, enum: [build_muscle, lose_fat, endurance, general_fitness] }
        fitness_level: { type: string, enum: [beginner, intermediate, advanced] }
        unit_system: { type: string, enum: [metric, imperial] }
        training_type: { type: string, enum: [running, strength, hybrid] }
        available_days:
          type: array
          items: { type: integer, minimum: 0, maximum: 6 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UserProfilePatch:
      type: object
      properties:
        name: { type: string }
        age: { type: integer }
        gender: { type: string, enum: [male, female, other] }
        height_cm: { type: number }
        weight_kg: { type: number }
        goal: { type: string, enum: [build_muscle, lose_fat, endurance, general_fitness] }
        fitness_level: { type: string, enum: [beginner, intermediate, advanced] }
    UserPreferencePatch:
      type: object
      properties:
        unit_system: { type: string, enum: [metric, imperial] }
        training_type: { type: string, enum: [running, strength, hybrid] }
        available_days:
          type: array
          items: { type: integer, minimum: 0, maximum: 6 }
...